requirements.txt
psycopg2-binary
bcrypt
pwinput
python-dotenv
pick
